move lto pipe into separate module
