+++ libs/common/src/lib/pipes/lto/index.ts
@@ -1,0 +1,2 @@
+export * from './lto-pipe.module';
+export * from './lto.pipe';
