+++ libs/common/src/lib/pipes/lto/lto-pipe.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { LtoPipe } from './lto.pipe';
+
+@NgModule({
+  declarations: [LtoPipe],
+  exports: [LtoPipe]
+})
+export class LtoPipeModule {}
